Fix looback device index in tests

+++ tests/finalize.sh
@@ -1,14 +1,18 @@
-echo 'Disable docker-lvmpy service'
+echo "Disable docker-lvmpy service"
 systemctl disable docker-lvmpy
-echo 'Removing all volumes from schain volume group'
+echo "Removing all volumes from schain volume group"
 lvremove schains --yes
-echo 'Removing volume group schain'
+echo "Removing volume group schain"
 vgremove schains
-echo 'Cleaning up /dev/loop0'
-pvremove /dev/loop0
-echo 'Unmount /dev/loop0'
-umount /dev/loop0
-echo 'Removing /dev/loop0'
-losetup -d /dev/loop0
-echo 'Removing loopbackfile.img'
-rm -f loopbackfile.img
+echo "Cleaning up $BLOCK_DEVICE"
+pvremove $BLOCK_DEVICE
+echo "Unmount $BLOCK_DEVICE"
+umount $BLOCK_DEVICE
+
+BLOCK_DEVICE="$(losetup --list -a | grep loopbackfile.img |  awk '{print $1}')"
+if [ ! -z ${BLOCK_DEVICE} ]; then
+    echo "Removing $BLOCK_DEVICE"
+    losetup -d $BLOCK_DEVICE
+    echo 'Removing loopbackfile.img'
+    rm -f loopbackfile.img
+fi
